src/pi/requirements.txt
pyserial 
keyboard